Created IL Autosplitter
Individual level autosplitter, starts on player taking control at the start of each level - splits when entering the hub. Intended for single split layouts.

One thing - if you are in the starting location of a level and reload a save from that area it wont start the timer due to the timer starting only when  starting in that location which it did you never left it you are still there.